+++ packages/edit/src/components/Edit.vue
@@ -1,15 +1,11 @@
 <template>
-  <VSheet class="tce-page-break pa-1" color="primary-darken-3">
-    Section break
-    <VIcon class="ml-4 pb-1" color="secondary-lighten-2" size="24">
-      mdi-format-page-break
-    </VIcon>
-  </VSheet>
+  <div class="tce-page-break">
+    <VDivider aria-label="Section break" opacity="0.5" variant="dashed">
+      <span
+        class="text-label-medium text-medium-emphasis text-uppercase font-weight-semibold"
+      >
+        Section break
+      </span>
+    </VDivider>
+  </div>
 </template>
-
-<style lang="scss" scoped>
-.tce-page-break {
-  text-align: center;
-  font-size: 1.125rem;
-}
-</style>